Rename local deploy functions, move file into local cli directory, modify test
