+++ chains/evm/cli/local/deploy.go
@@ -1,7 +1,7 @@
 // Copyright 2021 ChainSafe Systems
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package evm
+package local
 
 import (
 	"fmt"
@@ -35,9 +35,9 @@ type E2EClient interface {
 	calls.ClientDeployer
 }
 
-func PrepareEVME2EEnv(ethClient E2EClient, fabric calls.TxFabric, domainID uint8, threshold *big.Int, mintTo common.Address) (common.Address, common.Address, common.Address, error) {
+func PrepareLocalEnv(ethClient E2EClient, fabric calls.TxFabric, domainID uint8, threshold *big.Int, mintTo common.Address) (common.Address, common.Address, common.Address, error) {
 	staticGasPricer := evmgaspricer.NewStaticGasPriceDeterminant(ethClient, nil)
-	bridgeAddr, erc20Addr, erc20HandlerAddr, err := deployForTest(ethClient, fabric, staticGasPricer, domainID, threshold)
+	bridgeAddr, erc20Addr, erc20HandlerAddr, err := deployForLocal(ethClient, fabric, staticGasPricer, domainID, threshold)
 	if err != nil {
 		return common.Address{}, common.Address{}, common.Address{}, err
 	}
@@ -95,7 +95,7 @@ func PrepareEVME2EEnv(ethClient E2EClient, fabric calls.TxFabric, domainID uint8
 	return bridgeAddr, erc20Addr, erc20HandlerAddr, nil
 }
 
-func deployForTest(c E2EClient, fabric calls.TxFabric, gasPriceClient calls.GasPricer, domainID uint8, threshold *big.Int) (common.Address, common.Address, common.Address, error) {
+func deployForLocal(c E2EClient, fabric calls.TxFabric, gasPriceClient calls.GasPricer, domainID uint8, threshold *big.Int) (common.Address, common.Address, common.Address, error) {
 	erc20Addr, err := calls.DeployErc20(c, fabric, gasPriceClient, "Test", "TST")
 	if err != nil {
 		return common.Address{}, common.Address{}, common.Address{}, fmt.Errorf("ERC20 deploy failed: %w", err)
